requirements.txt
# Core Django
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.7
redis==5.0.1

# API Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3

# Real-time features
channels==4.0.0
channels-redis==4.1.0

# File handling
Pillow==10.0.1
django-storages==1.14.2
boto3==1.29.7

# Security & Authentication
django-allauth==0.57.0
django-guardian==2.4.0
cryptography==41.0.7

# Monitoring & Logging
sentry-sdk==1.38.0
django-debug-toolbar==4.2.0

# Performance
django-cachalot==2.6.1
gunicorn==21.2.0
whitenoise==6.6.0

# Development tools
black==23.10.0
isort==5.12.0
flake8==6.1.0
pytest-django==4.7.0

# AI/ML features (optional)
openai==1.3.0
celery==5.3.4
